Added new sitl mode Various renamings regarding unit tests Work on SITL mode drivers

+++ hc3d-tm/hc3d-tm/src/drivers/driver_tach.h
@@ -11,4 +11,4 @@
 #include "stdint.h"
 
 void driver_tach_init(void);
-uint32_t driver_tach_get(void);
+uint8_t driver_tach_get(void);
